many changes. add cloudinary for images also.

+++ backend/Controller/productController.js
@@ -1,12 +1,33 @@
-import mongoose from "mongoose";
 import productModel from "../Models/productModel.js";
 import fs from 'fs';
 import path from 'path';
+import cloudinary from "../utils/cloudinaryConfig.js";
+import { fileURLToPath } from 'url';
+
+// Define __dirname
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 export const addProduct = async(req,res)=>{
     try {
         const {title,description,originalprice,sellingprice,discount,category,brandname,weight,type} = req.body
-        const images = req.files.map(file => file.filename);
+
+        const images = [];
+
+        for(const file of req.files){
+            try {
+                const result = await cloudinary.uploader.upload(file.path,{folder: 'products'})
+                images.push({
+                    url: result.secure_url,
+                    public_id: result.public_id
+                })
+                fs.unlinkSync(path.join(__dirname, '../uploads/products', file.filename))
+            } catch (error) {
+                console.log('Error uploading file:', file.filename, error)
+            }
+        }
+
+
         const newProduct = new productModel({title,images,description,originalprice,sellingprice,discount,category,brandname,weight,type})
         await newProduct.save()
         res.status(200).json({message:"New prodect added", newProduct})
@@ -45,15 +66,27 @@ export const deleteProduct = async(req,res)=>{
 
     try {
         const product = await productModel.findOne({_id: id})
+        if (!product) {
+            return res.status(404).json({ message: "Product not found" });
+        }
         const images = product.images;
-        await productModel.deleteOne({_id: id})
 
-        images.forEach((image)=>{
-            const filePath = path.join(process.cwd(), 'uploads/products', image);
-            fs.unlink(filePath, (err) => {
-                if (err) console.log(`Error deleting image: ${image}`, err);
-              });
-        })
+        for(const image of images){
+            try {
+                await cloudinary.uploader.destroy(image.public_id)
+                console.log(`Deleted image from Cloudinary: ${image.public_id}`);
+            } catch (error) {
+                console.log(`Error deleting image from Cloudinary: ${image.public_id}`, error);
+            }
+        }
+
+        // images.forEach((image)=>{
+        //     const filePath = path.join(process.cwd(), 'uploads/products', image);
+        //     fs.unlink(filePath, (err) => {
+        //         if (err) console.log(`Error deleting image: ${image}`, err);
+        //       });
+        // })
+        await productModel.deleteOne({_id: id})
         res.status(200).json({message: "Product deleted successfully"})
     } catch (error) {
         res.status(500).json({message: "Problem in delete product"})
